feat: use environment everywhere for env vars
