+++ packages/substream/index.ts
@@ -1,15 +1,14 @@
 import { Command } from 'commander';
-import dotenv from 'dotenv';
-import { Duration, Effect, Either, Layer, Schedule, pipe } from 'effect';
+import { Duration, Effect, Either, Schedule, pipe } from 'effect';
 
 import { bootstrapRoot } from './sink/bootstrap-root';
 import { Environment, EnvironmentLive } from './sink/environment';
 import { getStreamConfiguration } from './sink/get-stream-configuration';
 import { runStream } from './sink/run-stream';
-import { Telemetry, TelemetryLive, startLogs } from './sink/telemetry';
+import { Telemetry, TelemetryLive } from './sink/telemetry';
 import { resetPublicTablesToGenesis } from './sink/utils/reset-public-tables-to-genesis';
 
-dotenv.config();
+function initialize() {}
 
 async function main() {
   const program = new Command();
@@ -21,16 +20,6 @@ async function main() {
   // @TODO: How do we make the options typesafe?
   const options = program.opts();
 
-  startLogs();
-
-  /**
-   * @TODO: It probably makes more sense to tie resetting the DB to a separate flag.
-   *        There are probably scenarios where we want to index from the genesis block
-   *        but not reset the DB.
-   *
-   *        I'd assume that `--from-genesis` starts from the genesis block and doesn't
-   *        have any side effects related to the DB.
-   */
   if (options.resetDb) {
     console.info('Resetting public tables');
     const reset = await pipe(resetPublicTablesToGenesis(), Effect.either, Effect.runPromise);
