--- packages/substream/sink/telemetry/index.ts
@@ -1,1 +1,0 @@
-export { Telemetry, TelemetryLive, startLogs } from './telemetry';
