--- packages/substream/sink/telemetry/telemetry.ts
@@ -1,48 +1,0 @@
-import * as Sentry from '@sentry/node';
-import { Context, Effect } from 'effect';
-
-export function startLogs() {
-  if (!process.env.TELEMETRY_URL) {
-    console.log('Telemetry environment not found. Skipping telemetry.');
-
-    return;
-  }
-
-  Sentry.init({
-    dsn: process.env.TELEMETRY_URL,
-
-    // We recommend adjusting this value in production, or using tracesSampler
-    // for finer control
-    tracesSampleRate: 1.0,
-  });
-}
-
-export function captureException(error: unknown) {
-  if (!process.env.TELEMETRY_URL) {
-    return;
-  }
-
-  Sentry.captureException(error);
-}
-
-export function captureMessage(message: string) {
-  if (!process.env.TELEMETRY_URL) {
-    return;
-  }
-
-  Sentry.captureMessage(message);
-}
-
-interface ITelemetry {
-  captureException: (error: unknown) => void;
-  captureMessage: (message: string) => void;
-}
-
-export class Telemetry extends Context.Tag('Telemetry')<Telemetry, ITelemetry>() {}
-
-export const TelemetryLive: ITelemetry = {
-  // Alternatively we can check for the presence of our Sentry env var and swap implementations here
-  // instead of having to read it in each capture method
-  captureException,
-  captureMessage,
-};
